Fix routing in table view
In some cases, a given AA resource won't have the show view defined,
like when you just want sub-resources to be reorderable. This commit
fixes that by looking directly at the route prefix on the AA resource
definition, which is usually the namespace that the resource lives
under. It then build the reorder URL that way.